--- backend/bottari/src/main/java/com/bottari/push/NotificationChannel.java
@@ -1,4 +1,0 @@
-package com.bottari.push;
-
-public interface NotificationChannel extends PushChannel {
-}
