+++ backend/bottari/src/main/java/com/bottari/push/connection/ConnectionChannel.java
@@ -1,5 +1,6 @@
-package com.bottari.push;
+package com.bottari.push.connection;
 
+import com.bottari.push.PushChannel;
 import com.bottari.push.message.PushMessage;
 
 public interface ConnectionChannel extends PushChannel {
