+++ backend/bottari/src/main/java/com/bottari/push/connection/ConnectionChannels.java
@@ -1,5 +1,7 @@
-package com.bottari.push;
+package com.bottari.push.connection;
 
+import com.bottari.push.ChannelType;
+import com.bottari.push.PushChannel;
 import com.bottari.push.message.PushMessage;
 import java.util.List;
 import java.util.Map;
