+++ backend/bottari/src/main/java/com/bottari/push/connection/sse/SseConnector.java
@@ -1,4 +1,4 @@
-package com.bottari.push.sse;
+package com.bottari.push.connection.sse;
 
 import com.bottari.config.MemberIdentifier;
 import lombok.RequiredArgsConstructor;
