+++ backend/bottari/src/main/java/com/bottari/push/notification/NotificationChannel.java
@@ -1,0 +1,6 @@
+package com.bottari.push.notification;
+
+import com.bottari.push.PushChannel;
+
+public interface NotificationChannel extends PushChannel {
+}
