+++ backend/bottari/src/test/java/com/bottari/fcm/FcmChannelTest.java
@@ -11,9 +11,9 @@ import com.bottari.error.BusinessException;
 import com.bottari.fixture.FcmTokenFixture;
 import com.bottari.fixture.MemberFixture;
 import com.bottari.member.domain.Member;
-import com.bottari.push.fcm.domain.FcmToken;
-import com.bottari.push.fcm.service.FcmChannel;
-import com.bottari.push.fcm.service.FcmTokenService;
+import com.bottari.push.notification.fcm.domain.FcmToken;
+import com.bottari.push.notification.fcm.service.FcmChannel;
+import com.bottari.push.notification.fcm.service.FcmTokenService;
 import com.bottari.push.message.MessageEventType;
 import com.bottari.push.message.MessageResourceType;
 import com.bottari.push.message.PushMessage;
