+++ backend/bottari/src/test/java/com/bottari/fixture/FcmTokenFixture.java
@@ -1,6 +1,6 @@
 package com.bottari.fixture;
 
-import com.bottari.push.fcm.domain.FcmToken;
+import com.bottari.push.notification.fcm.domain.FcmToken;
 import com.bottari.member.domain.Member;
 
 public enum FcmTokenFixture {
