+++ backend/bottari/src/test/java/com/bottari/member/service/MemberServiceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.bottari.error.BusinessException;
-import com.bottari.push.fcm.domain.FcmToken;
+import com.bottari.push.notification.fcm.domain.FcmToken;
 import com.bottari.member.domain.Member;
 import com.bottari.member.dto.CheckRegistrationResponse;
 import com.bottari.member.dto.CreateMemberRequest;
